Refactorizando clase service con dtos

+++ sav-msvc-mascota/src/main/java/com/jr/sav_msvc_patient_admission/services/PetServiceImpl.java
@@ -1,43 +1,67 @@
 package com.jr.sav_msvc_patient_admission.services;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
-
+import com.jr.sav_msvc_patient_admission.dto.PetDto;
+import com.jr.sav_msvc_patient_admission.mapper.PetMapper;
 import com.jr.sav_msvc_patient_admission.models.Pet;
 import com.jr.sav_msvc_patient_admission.repositories.PetsRepository;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class PetServiceImpl implements PetService{
 
     private final PetsRepository petsRepository;
+    private final PetMapper petMapper;
 
-    public PetServiceImpl(PetsRepository petsRepository) {
+    public PetServiceImpl(PetsRepository petsRepository, PetMapper petMapper) {
         this.petsRepository = petsRepository;
+        this.petMapper = petMapper;
     }
 
     @Override
-    public Pet savePet(Pet pet){
-        return petsRepository.save(pet);
+    public Optional<Pet> savePet(PetDto petDto){
+        if (petsRepository.findByNameAndOwnerNumber(petDto.getName(), petDto.getOwnerNumber()).isPresent()) {
+            return Optional.empty();
+        }
+        Pet entity = petMapper.toEntity(petDto);
+        petDto.setDateOfRecording(LocalDate.now());
+        return Optional.of(petsRepository.save(entity));
     }
 
     @Override
-    public List<Pet> findAllPets(){
-        return (List<Pet>) petsRepository.findAll();
+    public List<PetDto> findAllPets(){
+        return petsRepository.findAll().stream()
+            .map(petMapper::toDto).toList();
     }
 
     @Override
-    public Optional<Pet> findPetById(Long idPet){
-        return petsRepository.findById(idPet);
+    public Optional<PetDto> findPetById(Long idPet){
+        Optional<Pet> optPet = petsRepository.findById(idPet);
+        if (optPet.isEmpty()) {
+            return Optional.empty();
+        }
+        PetDto dto = petMapper.toDto(optPet.get());
+        return Optional.of(dto);
     }
 
     @Override 
-    public Optional<Pet> findByNameAndOwerId(String name, Long owerId){
-        return petsRepository.findByNameAndOwerId(name, owerId);
+    public Optional<PetDto> findByNameAndOwnerNumber(String name, Long ownerNumber){
+        Optional<Pet> optPet = petsRepository.findByNameAndOwnerNumber(name, ownerNumber);
+        if (optPet.isEmpty()) {
+            return Optional.empty();
+        }
+        PetDto dto = petMapper.toDto(optPet.get());
+        return Optional.of(dto);
     }
 
     @Override
     public void deletePetById(Long idPet){
+        if (petsRepository.findById(idPet).isEmpty()) {
+            throw new EntityNotFoundException();
+        }
         petsRepository.deleteById(idPet);
     }
 }
